修改店铺详情的信息，完成根据店铺id查询店铺商品详情列表。

+++ src/main/java/com/pro516/thrifttogetherbackstage/entity/vo/CollectedProductVO.java
@@ -26,7 +26,7 @@ public class CollectedProductVO {
     private String coverUrl;
 
     @ApiModelProperty(value = "商品价格")
-    private Integer productPrice;
+    private Double productPrice;
 
     @ApiModelProperty(value = "店铺名称")
     private String shopName;
